refactor: move Admin table backend, no testing yet
