+++ src/lib/classes/Admin.ts
@@ -1,17 +1,12 @@
-import {
-	insertAdminDB,
-	selectAdminDB,
-	updateAdminDB,
-	deleteAdminDB
-} from '$lib/server/AdminSB';
+import { insertAdminDB, selectAdminDB, updateAdminDB, deleteAdminDB } from '$lib/server/AdminSB';
 import type { AdminProcessed } from '$lib/utils/types';
 
 // parameter type for insert and update admin DB functions
 export type AdminDBObj = {
 	admin_id: number;
 	rfid: number;
-    pw: string;
-    nickname: string;
+	pw: string;
+	nickname: string;
 };
 
 // return value of admin DB functions
@@ -24,26 +19,26 @@ export type AdminResponse = {
 // filters for selecting admin records
 export type AdminFilter = {
 	adminID: number;
-    nickname: string;
+	nickname: string;
 };
 
 export class Admin {
 	/* Contains all admin methods for conversion and DB communication. */
 
 	public static toAdminDBObj(admin: AdminProcessed): AdminDBObj {
-		/* Converts a AdminProcessed to a AdminDBObj. */
+		/* Converts aa AdminProcessed to aa AdminDBObj. */
 		return {
 			admin_id: admin.adminID,
-            rfid: 0,
-            pw: '',
-            nickname: admin.nickname
+			rfid: 0,
+			pw: '',
+			nickname: 'nickname' in admin ? admin.nickname : ''
 		};
 	}
 
 	public static async selectAdmins(
 		filter: AdminFilter = {
-            adminID: 0,
-            nickname: ''
+			adminID: 0,
+			nickname: ''
 		}
 	): Promise<AdminResponse> {
 		/* Selects all admin records in database using the default or given filter. */
