+++ src/routes/dashboard/admin/admins/+page.server.ts
@@ -1,0 +1,6 @@
+import { Admin } from '$lib/classes/Admin';
+
+export async function load() {
+	/* Loads admin records from the DB when page is created. */
+	return Admin.selectAdmins();
+}
